Run checkpatch on entire files
This changes the approach so that the hook can be run the same in CI and
local development by running on whole files by using the --file argument
of the checkpatch.pl script.

This change means that the entire file must be compliant with configured
checks instead of just the lines that are being changed, and so
precludes usage whereby old projects can be made compliant in a
progressive, piecemeal manner.